requirements.txt
asgiref==3.3.1
Django==3.1.3
pytz==2020.4
sqlparse==0.4.1
requests==2.25.1
numpy
sklearn
django-cors-headers==3.10.1
djangorestframework==3.13.1
